llm/requirements.txt
# Dependencias para sistema de embeddings y Pinecone
pinecone-client>=3.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Optimizaciones opcionales
# optimum[onnxruntime]>=1.15.0  # Para optimización ONNX
# optimum-intel[openvino]>=1.15.0  # Para optimización OpenVINO 